RavenDB-11919 Fix DistributedRevisionsSubscription test
We have to wait for the subscription to be created on the responsible node.
